added initialization of k means method

+++ dataset_read.h
@@ -19,7 +19,7 @@ int ReverseInt(int i)
 	ch4 = (i >> 24) & 255;
 	return((int)ch1 << 24) + ((int)ch2 << 16) + ((int)ch3 << 8) + ch4;
 }
-void ReadMNIST(string path, int NumberOfImages, int DataOfAnImage, thrust::host_vector<short> &arr)
+void ReadMNIST(string path, int NumberOfImages, int DataOfAnImage, thrust::host_vector<float> &arr)
 {
 	arr.resize(NumberOfImages*DataOfAnImage);
 	std::
@@ -47,7 +47,7 @@ void ReadMNIST(string path, int NumberOfImages, int DataOfAnImage, thrust::host_
 				{
 					unsigned char temp = 0;
 					file.read((char*)&temp, sizeof(temp));
-					arr[(ll)(i*(DataOfAnImage)+(n_cols*r) + c)] = (short)temp;
+					arr[(ll)(i*(DataOfAnImage)+(n_cols*r) + c)] = (float)temp;
 				}
 			}
 		}
